chore: rename files to match all other test files
