+++ tests/e2e/auth/signup/sign-up-page-fixture.ts
